requirements.txt
azure-functions
azure-storage-blob
sharepy
pandas==1.0.5
xlsxwriter
xlrd
seaborn
jinja2